refactor: move nearby-stop-places into screen-components
