+++ src/screen-components/nearby-stop-places/components/StopPlaceItem.tsx
