+++ src/screen-components/nearby-stop-places/components/StopPlaces.tsx
