+++ src/screen-components/nearby-stop-places/index.tsx
@@ -1,2 +1,3 @@
 export {NearbyStopPlacesScreenComponent} from './NearbyStopPlacesScreenComponent';
+export type {NearbyStopPlacesScreenParams} from './NearbyStopPlacesScreenComponent';
 export type {StopPlacesMode} from './types';
