+++ src/screen-components/nearby-stop-places/types.ts
